+++ astropy/coordinates/angles.py
@@ -651,20 +651,36 @@ class Longitude(Angle):
         return obj
 
 #<----------------------------------Rotations--------------------------------->
+# The main routines have been moved to matrix_utilities.  The definitions here
+# are for backward compatibility.
+from . import matrix_utilities
+from ..utils.decorators import deprecated
+
+_DEPRECATION_MESSAGE="""
+Numpy matrix instances are no longer used to represent rotation matrices, since
+they do not allow one to represent stacks of matrices. Instead, plain arrays
+are used. Instead of %(func)s, use %(alternative)s.
+For matrix multiplication and tranposes, it is suggested to use
+:func:`~astropy.coordinates.matrix_utilities.matrix_product` and
+:func:`~astropy.coordinates.matrix_utilities.matrix_transpose`, respectively.
+"""
+@deprecated(since='1.3', message=_DEPRECATION_MESSAGE, alternative=':func:'
+            '`~astropy.coordinates.matrix_utilities.rotation_matrix`')
 def rotation_matrix(angle, axis='z', unit=None):
     """
-    Generate rotation matrices for rotation by some angle(s).
+    Generate a 3x3 cartesian rotation matrix in for rotation about
+    a particular axis.
 
     Parameters
     ----------
     angle : convertible to `Angle`
-        The amount of rotation the matrices should represent.
+        The amount of rotation this matrix should represent.
 
-    axis : str, or array-like
-        Either ``'x'``, ``'y'``, ``'z'``, or a (x,y,z) specifying the axis to
-        rotate about. If ``'x'``, ``'y'``, or ``'z'``, the rotation sense is
-        counterclockwise looking down the + axis (e.g. positive rotations obey
-        left-hand-rule).  If given as an array, the last dimension should be 3.
+    axis : str or 3-sequence
+        Either ``'x'``, ``'y'``, ``'z'``, or a (x,y,z) specifying an
+        axis to rotate about. If ``'x'``, ``'y'``, or ``'z'``, the
+        rotation sense is counterclockwise looking down the + axis
+        (e.g. positive rotations obey left-hand-rule).
 
     unit : UnitBase, optional
         If ``angle`` does not have associated units, they are in this
@@ -675,46 +691,15 @@ def rotation_matrix(angle, axis='z', unit=None):
     rmat: `numpy.matrix`
         A unitary rotation matrix.
     """
-    if unit is None:
-        unit = u.degree
-
-    angle = Angle(angle, unit=unit)
-
-    s = np.sin(angle)
-    c = np.cos(angle)
-
-    # use optimized implementations for x/y/z
-    try:
-        i = 'xyz'.index(axis)
-    except TypeError:
-        axis = np.asarray(axis)
-        axis = axis / np.sqrt((axis * axis).sum(axis=-1, keepdims=True))
-        R = (axis[..., np.newaxis] * axis[..., np.newaxis, :] *
-             (1. - c)[..., np.newaxis, np.newaxis])
-
-        for i in range(0, 3):
-            R[..., i, i] += c
-            a1 = (i + 1) % 3
-            a2 = (i + 2) % 3
-            R[..., a1, a2] += axis[..., i] * s
-            R[..., a2, a1] -= axis[..., i] * s
-
-    else:
-        a1 = (i + 1) % 3
-        a2 = (i + 2) % 3
-        R = np.zeros(angle.shape + (3, 3))
-        R[..., i, i] = 1.
-        R[..., a1, a1] = c
-        R[..., a1, a2] = s
-        R[..., a2, a1] = -s
-        R[..., a2, a2] = c
-
-    return R
+    return matrix_utilities.rotation_matrix(angle, axis, unit).view(np.matrix)
 
 
+@deprecated(since='1.3', message=_DEPRECATION_MESSAGE, alternative=':func:'
+            '`~astropy.coordinates.matrix_utilities.angle_axis`')
 def angle_axis(matrix):
     """
-    Angle of rotation and rotation axis for a given rotation matrix.
+    Computes the angle of rotation and the rotation axis for a given rotation
+    matrix.
 
     Parameters
     ----------
@@ -729,15 +714,5 @@ def angle_axis(matrix):
     axis : array (length 3)
         The (normalized) axis of rotation for this matrix.
     """
-    m = np.asanyarray(matrix)
-    if m.shape[-2:] != (3, 3):
-        raise ValueError('matrix is not 3x3')
-
-    axis = np.zeros(m.shape[:-1])
-    axis[..., 0] = m[..., 2, 1] - m[..., 1, 2]
-    axis[..., 1] = m[..., 0, 2] - m[..., 2, 0]
-    axis[..., 2] = m[..., 1, 0] - m[..., 0, 1]
-    r = np.sqrt((axis * axis).sum(-1, keepdims=True))
-    angle = np.arctan2(r[..., 0],
-                       m[..., 0, 0] + m[..., 1, 1] + m[..., 2, 2] - 1.)
-    return Angle(angle, u.radian), -axis / r
+    m = np.asmatrix(matrix)
+    return matrix_utilities.angle_axis(m.view(np.ndarray))
