+++ astropy/coordinates/representation.py
@@ -368,7 +368,7 @@ class CartesianRepresentation(BaseRepresentation):
 
         We now create a rotation matrix around the z axis:
 
-            >>> from astropy.coordinates.angles import rotation_matrix
+            >>> from astropy.coordinates.matrix_utilities import rotation_matrix
             >>> rotation = rotation_matrix(30 * u.deg, axis='z')
 
         Finally, we can apply this transformation:
